Glue code for MMode reconciliation by DRCluster reconciler
- Add watches for MCV and MW
- Add RBAC to klusterlet for MMode CR management
- Update roles required for DRluster reconciler
- Update test cases as MW counts changed
- Garbage collect MMode resource on DRCluster deletion

+++ api/v1alpha1/maintenancemode_types.go
@@ -57,7 +57,7 @@ const (
 // MaintenanceModeStatus defines the observed state of MaintenanceMode
 type MaintenanceModeStatus struct {
 	State              MModeState         `json:"state,omitempty"`
-	ObservedGeneration int                `json:"observedGeneration,omitempty"`
+	ObservedGeneration int64              `json:"observedGeneration,omitempty"`
 	Conditions         []metav1.Condition `json:"conditions,omitempty"`
 }
